CLI Improvements for populate named ranges
Modified xl_populate_named_ranges.py to improve the user experience.
Rather than manually entering file names in the python script, the user
can now select from open workbooks that are seen by `xlwings`, and JSON
files in the users current working directory.

Additionally, the named ranges to be updated are listed with the old and
new proposed values, as well as the percent that these values change.
The user is prompted to confirm prior to overwriting values. The intent
of this update is to eliminate errors & go-backs.