Added file picker for nx_get_measurements
Used some tkinter hackery to get a file picker window to open in order
to allow the user to choose where to save the JSON file of measurements.
Defaults to C:\Users\<current
username>\Documents\datum\nx_measurements.json. Removes the need to
manually edit the Python journal if a different directory is desired.

Wrote up current justification for except/pass statement when reading
measurement values. Took some investigation to find out why I did it,
but I identified that this script only works for measurements with a
single numerical value. Other measurements types, primiarily points,
lists, and vectors, are not handled and are currently simply skipped. I
removed the `pass` statement and replaced it with a warning.

Also removed the `csvwrite` functionality as it was no longer needed or
baselined. Some other commenting & cleanup throughout.