Added Console Functionality
Added a module `datum_console.py` that uses a JSON configuration file to
select functions and generates a help message based on the config file.
Needs some work to be able to implement with XLPNR.

Added the import to XLPNR but did not implement.

Updated .gitignore to include `console_config.json`.