Updated console safety & funcationality
After reading about the dangers of using `eval()`, I re-did the console
module so that `eval()` is no longer used. Console commands are now
listed in the code, rather than within the JSON configuration file.
Rather than writing the documentation for the help message in the config
file, it simply uses the `function.__doc__` property (or the
`function.__name__` property if no docstring).

The JSON config file is now only used for high-level options of the
console, and not for listing commands - that stays in the code.

Also added functionality for arguments, which should make implementation
with XLPNR easier and more useful.