Simplified console.
Removed configuration file for console, and simply implemented the
defaults into the `datum_console.py` file. I had already removed the
commands and the docstrings from the config file, and the other options
were just extras, but didn't really add much value. It was a good exercise to perform. There is still a lot to
add to DATUM, but removal of un-needed features can be equally as
valuable.