NX Get Measurements Updates
Added two significant features to `nx_get_measurements.py`:
1) Export metadata in JSON along with measurements. Metadata includes
part name, revision, extraction date, user, and part units.
2) Support for points, vectors, strings, and list expression types.
These are now included in the JSON that's exported; however, there is no
implementation for them elsewhere in DATUM at this time. In addition to
this, the WCS feature is also exported, in order to give reference for
the MOI and POI that come out of mass measurements.

Updated xlpnr to change `expr["type"]` to `expr["description"]` since
`Expression.Type` is already used in the NX API (returns Point, Vector,
Number, etc.) and "description" now points to the end of the string that
tells us that this is a radius, mass, surface normal, etc.